english-assistant(chuban)/requirements.txt
# ============================================
# 英语学习助手 - 依赖列表
# ============================================
# 项目：智能英语发音评估与学习助手系统
# Python版本要求：>=3.8
# 最后更新：2025-08-26
# ============================================

# =====================================
# 核心Web框架
# =====================================
Flask==3.0.3
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# =====================================
# 数据库支持
# =====================================
# MySQL 数据库驱动
PyMySQL==1.1.0
mysqlclient==2.2.4

# ORM 框架
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1

# 数据库迁移
Flask-Migrate==4.0.5
Alembic==1.13.1

# 连接池
SQLAlchemy-Utils==0.41.1

# =====================================
# 用户认证和安全
# =====================================
PyJWT==2.8.0
bcrypt==4.1.2

# =====================================
# 深度学习和AI模型支持
# =====================================
# PyTorch生态系统（音素级评分核心依赖）
torch==2.2.2
torchaudio==2.2.2
transformers==4.41.2

# Hugging Face模型加载
huggingface-hub==0.21.4
tokenizers==0.15.2
safetensors==0.4.2

# =====================================
# 音频处理和分析
# =====================================
# 核心音频处理库
librosa==0.10.2.post1
soundfile==0.12.1
audioread==3.0.1

# 音频录制和播放
sounddevice==0.4.6
PyAudio==0.2.14

# 音频格式转换
ffmpeg-python==0.2.0

# =====================================
# 科学计算和数学库
# =====================================
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.4.0

# =====================================
# 数据处理和分析
# =====================================
pandas==2.2.1

# =====================================
# 自然语言处理
# =====================================
# 语法检查
language_tool_python==2.7.1

# 文本处理
nltk==3.8.1

# 文本相似度计算
python-Levenshtein==0.25.1
fuzzywuzzy==0.18.0

# =====================================
# 语音识别（备用方案）
# =====================================
SpeechRecognition==3.10.0

# =====================================
# 配置和工具
# =====================================
# 配置文件处理
PyYAML==6.0.1

# HTTP请求
requests==2.31.0

# 进度条
tqdm==4.66.1

# =====================================
# 开发和调试工具（可选）
# =====================================
# 音频可视化（可选）
# matplotlib==3.8.2
# seaborn==0.13.0

# 代码格式化（开发用）
# black==24.2.0
# flake8==7.0.0

# 测试框架（开发用）
# pytest==8.0.0
# pytest-cov==4.0.0

# =====================================
# 系统和平台特定依赖
# =====================================
# Windows平台音频支持
# pywin32==306; platform_system=="Windows"

# =====================================
# 版本兼容性说明
# =====================================
# 1. torch和transformers版本必须匹配
# 2. librosa需要ffmpeg支持
# 3. PyAudio在某些系统上需要手动编译
# 4. 推荐使用Python 3.8-3.11
# 5. 内存需求：4GB+（模型加载）

# =====================================
# 安装说明
# =====================================
# 基础安装：pip install -r requirements.txt
# 如果PyAudio安装失败：
#   Windows: pip install pipwin && pipwin install pyaudio
#   macOS: brew install portaudio && pip install pyaudio
#   Ubuntu: sudo apt-get install python3-pyaudio
#
# 如果torch安装失败，请访问：
#   https://pytorch.org/get-started/locally/
# =====================================